+++ back/src/forms/resolvers/mutations/createFormRevisionRequest.ts
@@ -1,4 +1,5 @@
 import {
+  BsddTransporter,
   EmitterType,
   Form,
   Prisma,
@@ -96,7 +97,10 @@ export default async function createFormRevisionRequest(
   const { formId, content, comment, authoringCompanySiret } = input;
 
   const user = checkIsAuthenticated(context);
-  const existingBsdd = await getFormOrFormNotFound({ id: formId });
+  const existingBsdd = await getFormOrFormNotFound(
+    { id: formId },
+    { transporters: true }
+  );
 
   const formRepository = getFormRepository(user);
 
@@ -211,7 +215,7 @@ async function checkIfUserCanRequestRevisionOnBsdd(
 
 async function getFlatContent(
   content: FormRevisionRequestContentInput,
-  bsdd: Form
+  bsdd: Form & { transporters: BsddTransporter[] }
 ): Promise<RevisionRequestContent> {
   const flatContent = flattenBsddRevisionRequestInput(content);
 
@@ -242,6 +246,26 @@ async function getFlatContent(
   ) {
     throw new ForbiddenError(
       "Impossible d'annuler un bordereau qui a été réceptionné sur l'installation de destination."
+    );
+  }
+
+  // If the BSD has been received, you can modify the quantityReceived. Else, no
+  if (content.quantityReceived && !bsdd.receivedAt) {
+    throw new ForbiddenError(
+      "Impossible de réviser la quantité reçue si le bordereau n'a pas encore été réceptionné."
+    );
+  }
+
+  // No more than 40 tons by ROAD transport
+  if (
+    content &&
+    content?.quantityReceived &&
+    content?.quantityReceived > 40 &&
+    bsdd?.transporters?.length === 1 &&
+    bsdd?.transporters[0]?.transporterTransportMode === "ROAD"
+  ) {
+    throw new ForbiddenError(
+      "La quantité reçue ne peut dépasser 40 tonnes pour le transporter routier."
     );
   }